Small updates of the cities of Spain
Hi there!

Thank you so much for sharing this great database of cities and countries! I will update on the later days the cities of Spain. I have also created the worlddb database and I was not able to find the city of Valencia in Spain (there is also a city with the same name in Venezuela, this is the city I found), I will try to find this problem and fix it.  :)

Thanks again!